+++ tests/testthat/test_popkin.R
@@ -146,11 +146,33 @@ test_that("function returns precomputed values: weights_subpops", {
 })
 
 test_that("function returns precomputed values: get_A", {
-    expect_equal(get_A(X), A)
-    expect_equal(get_A(X+0), A) # turns numeric
-    expect_equal(get_A(2L-X), A)
-    expect_equal(get_A(2-X), A) # numeric version again
+    # standard test (X is integer)
+    expect_silent( obj <- get_A( X ) )
+    expect_equal( obj$A, A )
+    expect_equal( obj$M, M )
+    
+    # turns X into doubles
+    expect_silent( obj <- get_A( X + 0 ) )
+    expect_equal( obj$A, A )
+    expect_equal( obj$M, M )
+
+    # reflect, keep X integer
+    expect_silent( obj <- get_A( 2L - X ) )
+    expect_equal( obj$A, A )
+    expect_equal( obj$M, M )
+
+    # reflect and turn X doubles too
+    expect_silent( obj <- get_A( 2 - X ) )
+    expect_equal( obj$A, A )
+    expect_equal( obj$M, M )
+
+    # these should be square matrices
     expect_equal(nrow(A), ncol(A))
+    expect_equal(nrow(M), ncol(M))
+
+    # M (pairwise sample sizes, so excluding NA pairs) must satisfy obvious range limits
+    expect_true( min(M) >= 0 )
+    expect_true( max(M) <= nrow(X) )
 })
 
 test_that("function returns precomputed values: min_mean_subpops", {
@@ -163,12 +185,26 @@ test_that("function returns precomputed values: min_mean_subpops", {
 # higher-level tests now!
 
 test_that("function returns precomputed values: popkin", {
+    # cause problems on purpose
+    expect_error( popkin() )
+    
+    # good runs
     expect_equal(popkin(X), Phi0)
     expect_equal(popkin(X, subpops0), Phi0)
     expect_equal(popkin(X, subpops), Phi)
     expect_equal(popkin(X+0, subpops), Phi)
     expect_equal(popkin(2L-X, subpops), Phi)
     expect_equal(popkin(2-X, subpops), Phi)
+
+    # test want_M in these two cases only
+    # though kinship is scaled differently in each version, M is the same for both
+    expect_silent( obj <- popkin( X, want_M = TRUE ) )
+    expect_equal( obj$kinship, Phi0 )
+    expect_equal( obj$M, M )
+    
+    expect_silent( obj <- popkin( X, subpops, want_M = TRUE ) )
+    expect_equal( obj$kinship, Phi )
+    expect_equal( obj$M, M )
 })
 
 test_that("popkin preserves names of individuals", {