+++ NEWS.md
@@ -211,9 +211,15 @@ More improvements to function `plot_popkin`:
 
 # 2020-07-13 - popkin 1.3.3.9000
 
-* Internal (non-exported) function `solve_m_mem_lim` now avoids a rare integer overflow caused when input number of inndividuals `n` was encoded as an integer and was greater than `sqrt(.Machine$integer.max)`, or 46340.95.
+* Internal (non-exported) function `solve_m_mem_lim` now avoids a rare integer overflow caused when input number of individuals `n` was encoded as an integer and was greater than `sqrt(.Machine$integer.max)`, or 46340.95.
 
 # 2020-07-23 - popkin 1.3.4.9000
 
 * `validate_kinship` now has `sym` option that, if `FALSE`, skips symmetry test (defaults to `TRUE`).
 * `plot_popkin` has the same `sym` option passed to `validate_kinship`, but here it defaults to `FALSE` (there is no inherent error caused by plotting non-symmetric matrices).
+
+# 2020-09-24 - popkin 1.3.5.9000
+
+* Function `popkin`
+  - Added `want_M` option, which if `TRUE` returns a list containing the `kinship` matrix as well as the pairwise complete count matrix `M`.
+  - Minor documentation improvements.